refactor: Melhorando nomenclatura da classe de validação e utilizando classe de comando Firely com lógica encapsulada ao invés de um método local, para melhor utilização do conceito de programação orientada a objetos
